fix: break command recovery import cycle (#116947)
